requirements.txt
django==1.5
vobject
django-cal
python-social-auth
django-piston
South
django-toolbelt